+++ Forge/src/main/java/com/nyfaria/spookybats/event/ClientModEvents.java
@@ -5,10 +5,7 @@ import com.nyfaria.spookybats.client.model.PumpkinBatModel;
 import com.nyfaria.spookybats.client.model.SkeletonBatModel;
 import com.nyfaria.spookybats.client.model.UndeadBatModel;
 import com.nyfaria.spookybats.client.model.WitchBatModel;
-import com.nyfaria.spookybats.client.renderer.CreeperBatRenderer;
-import com.nyfaria.spookybats.client.renderer.EmissiveBatRenderer;
-import com.nyfaria.spookybats.client.renderer.PlayerBatRenderer;
-import com.nyfaria.spookybats.client.renderer.VoidBatRenderer;
+import com.nyfaria.spookybats.client.renderer.*;
 import com.nyfaria.spookybats.client.renderer.api.SpookyBatRenderer;
 import com.nyfaria.spookybats.client.renderer.layer.BatWingsLayer;
 import com.nyfaria.spookybats.init.EntityInit;
@@ -24,7 +21,8 @@ public class ClientModEvents {
 
     @SubscribeEvent
     public static void onEntityRenderers(EntityRenderersEvent.RegisterRenderers event) {
-        event.registerEntityRenderer(EntityInit.PUMPKIN_BAT.get(), context -> new EmissiveBatRenderer(context, new PumpkinBatModel<>(context.bakeLayer(PumpkinBatModel.LAYER_LOCATION)), "pumpkin_bat"));
+        event.registerEntityRenderer(EntityInit.PUMPKIN_BAT.get(), context -> new MascotBatRenderer(context, new PumpkinBatModel<>(context.bakeLayer(PumpkinBatModel.LAYER_LOCATION)), "pumpkin_bat"));
+        event.registerEntityRenderer(EntityInit.UNDEAD_BAT.get(), context -> new MascotBatRenderer(context, new UndeadBatModel<>(context.bakeLayer(UndeadBatModel.LAYER_LOCATION)), "undead_bat"));
         event.registerEntityRenderer(EntityInit.CREEPER_BAT.get(), CreeperBatRenderer::new);
         event.registerEntityRenderer(EntityInit.WITCH_BAT.get(), context -> new SpookyBatRenderer<>(context, new WitchBatModel<>(context.bakeLayer(WitchBatModel.LAYER_LOCATION))));
         event.registerEntityRenderer(EntityInit.STEVE_BAT.get(), context -> new SpookyBatRenderer<>(context, new HatBatModel<>(context.bakeLayer(HatBatModel.LAYER_LOCATION))));
@@ -32,7 +30,6 @@ public class ClientModEvents {
         event.registerEntityRenderer(EntityInit.HEROBRINE_BAT.get(), context -> new EmissiveBatRenderer(context, new HatBatModel<>(context.bakeLayer(HatBatModel.LAYER_LOCATION)), "herobrine_bat"));
         event.registerEntityRenderer(EntityInit.SKELETON_BAT.get(), context -> new SpookyBatRenderer<>(context, new SkeletonBatModel<>(context.bakeLayer(SkeletonBatModel.LAYER_LOCATION))));
         event.registerEntityRenderer(EntityInit.WITHER_SKELETON_BAT.get(), context -> new SpookyBatRenderer<>(context, new SkeletonBatModel<>(context.bakeLayer(SkeletonBatModel.LAYER_LOCATION))));
-        event.registerEntityRenderer(EntityInit.UNDEAD_BAT.get(), context -> new EmissiveBatRenderer(context, new UndeadBatModel<>(context.bakeLayer(UndeadBatModel.LAYER_LOCATION)), "undead_bat"));
         event.registerEntityRenderer(EntityInit.PLAYER_BAT.get(), context -> new PlayerBatRenderer(context, new HatBatModel<>(context.bakeLayer(HatBatModel.LAYER_LOCATION))));
         event.registerEntityRenderer(EntityInit.VOID_BAT.get(), context -> new VoidBatRenderer(context, new HatBatModel<>(context.bakeLayer(HatBatModel.LAYER_LOCATION))));